make health check a periodic task
