+++ catalog/celery.py
@@ -1,5 +1,6 @@
 import os
 
+from django.conf import settings
 from celery import Celery
 
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "catalog.settings")
@@ -8,7 +9,11 @@ app = Celery("catalog")
 app.config_from_object("django.conf:settings", namespace="CELERY")
 app.autodiscover_tasks()
 
-
-@app.task(bind=True)
-def debug_task(self):
-    print(f"Request: {self.request!r}")
+app.conf.timezone = settings.TIME_ZONE
+app.conf.beat_schedule = {
+    'send-active-checks-to-github': {
+        'task': 'health.tasks.send_active_to_github',
+        'schedule': 60 * 5, # Every 5 minutes.
+        'args': (os.environ.get("CRON_USER"),),
+    }
+}
