+++ catalog/supervisord.conf
@@ -24,7 +24,7 @@ stdout_logfile_maxbytes = 0
 
 [program:celery]
 directory=/code
-command=celery -A catalog worker -l info
+command=celery -A catalog worker -l ERROR -B
 stdout_logfile=/dev/stdout
 stderr_logfile=/dev/stderr
 stdout_logfile_maxbytes = 0
